NCKH/requirements.txt
ipykernel
python-docx
chromadb
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-llms-groq

